moved alignment to nwbconverter
